requirements.txt
torch
torchvision
Pillow
requests
beautifulsoup4
pymupdf
matplotlib
numpy
streamlit
